# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Papertrail API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/brands": {
      "get": {
        "tags": [
          "brands"
        ],
        "summary": "List Brands",
        "description": "Every brand with how many archived receipts it matches, plus the unmatched ones.",
        "operationId": "list_brands_api_brands_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandsOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "brands"
        ],
        "summary": "Create Brand",
        "operationId": "create_brand_api_brands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/brands/{brand_id}": {
      "patch": {
        "tags": [
          "brands"
        ],
        "summary": "Update Brand",
        "description": "Rename a brand or replace its prefixes. The id never changes, so receipts keep their grouping.",
        "operationId": "update_brand_api_brands__brand_id__patch",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "brands"
        ],
        "summary": "Delete Brand",
        "description": "Remove a brand. Its receipts stay where they are; they simply stop being grouped.",
        "operationId": "delete_brand_api_brands__brand_id__delete",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/brands/{brand_id}/prefixes": {
      "post": {
        "tags": [
          "brands"
        ],
        "summary": "Add Prefix",
        "description": "Add one prefix — the quick action on a suggestion.",
        "operationId": "add_prefix_api_brands__brand_id__prefixes_post",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandPrefixIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/brands/suggestions": {
      "get": {
        "tags": [
          "brands"
        ],
        "summary": "Suggestions",
        "description": "Prefixes shared by unmatched receipt names. The settings are remembered, as in Streamlit.",
        "operationId": "suggestions_api_brands_suggestions_get",
        "parameters": [
          {
            "name": "boundary_only",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Boundary Only"
            }
          },
          {
            "name": "max_length",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 80,
                  "minimum": 4
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Length"
            }
          },
          {
            "name": "min_length",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 24,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Length"
            }
          },
          {
            "name": "min_count",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 50,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrefixSuggestionOut"
                  },
                  "title": "Response Suggestions Api Brands Suggestions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrandBranchOut": {
        "properties": {
          "location": {
            "type": "string",
            "title": "Location"
          },
          "receipts": {
            "type": "integer",
            "title": "Receipts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "location",
          "receipts"
        ],
        "title": "BrandBranchOut"
      },
      "BrandIn": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "prefixes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prefixes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "prefixes"
        ],
        "title": "BrandIn"
      },
      "BrandOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "prefixes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prefixes"
          },
          "receipt_count": {
            "type": "integer",
            "title": "Receipt Count"
          },
          "tree": {
            "items": {
              "$ref": "#/components/schemas/BrandPrefixOut"
            },
            "type": "array",
            "title": "Tree"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "prefixes",
          "receipt_count",
          "tree"
        ],
        "title": "BrandOut"
      },
      "BrandOverview": {
        "properties": {
          "receipts": {
            "type": "integer",
            "title": "Receipts"
          },
          "matched": {
            "type": "integer",
            "title": "Matched"
          },
          "unmatched": {
            "type": "integer",
            "title": "Unmatched"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "receipts",
          "matched",
          "unmatched"
        ],
        "title": "BrandOverview"
      },
      "BrandPrefixIn": {
        "properties": {
          "prefix": {
            "type": "string",
            "title": "Prefix"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prefix"
        ],
        "title": "BrandPrefixIn"
      },
      "BrandPrefixOut": {
        "properties": {
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "receipts": {
            "type": "integer",
            "title": "Receipts"
          },
          "branches": {
            "items": {
              "$ref": "#/components/schemas/BrandBranchOut"
            },
            "type": "array",
            "title": "Branches"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prefix",
          "receipts",
          "branches"
        ],
        "title": "BrandPrefixOut"
      },
...
      "BrandsOut": {
        "properties": {
          "brands": {
            "items": {
              "$ref": "#/components/schemas/BrandOut"
            },
            "type": "array",
            "title": "Brands"
          },
          "overview": {
            "$ref": "#/components/schemas/BrandOverview"
          },
          "unmatched": {
            "items": {
              "$ref": "#/components/schemas/NameCount"
            },
            "type": "array",
            "title": "Unmatched"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "brands",
          "overview",
          "unmatched"
        ],
        "title": "BrandsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NameCount": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "count"
        ],
        "title": "NameCount"
      },
...
      "PrefixSuggestionOut": {
        "properties": {
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prefix",
          "count",
          "names"
        ],
        "title": "PrefixSuggestionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}